# Film Manager Implementation/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Film Manager",
    "description": "REST APIs for the interaction with the Film Manager service.",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
...
  "paths": {
...
    "/api/films/public": {
      "get": {
        "tags": [
          "api/films/public"
        ],
        "summary": "Retrieve the public films",
        "description": "The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is used to limit the size of messages.",
        "operationId": "getPublicFilms",
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "The id of the requested page (if absent, the first page is returned)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Films"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
...
    "/api/films/public/{filmId}/reviews": {
      "post": {
        "tags": [
          "api/films/public/{filmId}/reviews"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Issue film review to some users",
        "description": "The film with ID filmId is assigned to one or more users for review and the corresponding reviews are created. The users are specified in the review representations in the request body. This operation can only be performed by the owner.",
        "operationId": "issueFilmReview",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "the new film reviews, including the users to whom they are issued",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Review"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (user is not the owner)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "409": {
            "description": "Conflict (e.g., wrong film id or unknown reviewer id)",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "get": {
        "tags": [
          "api/films/public/{filmId}/reviews"
        ],
        "summary": "Retrieve the list of all the reviews that have been issued/completed for a public film",
        "description": "All the reviews that have been issued/completed for the public film with ID filmId are retrieved. A pagination mechanism is used to limit the size of messages. This operation does not require authentication. ",
        "operationId": "getFilmReviews",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose reviews must be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "name": "pageNo",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the requested page (if absent, the first page is returned)'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reviews"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
    "/api/films/public/{filmId}/reviews/{reviewerId}": {
      "get": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}"
        ],
        "summary": "Retrieve a review that has been issued/completed for a film",
        "description": "The review of the film with ID filmID issued to the user with ID reviewerId is retrieved. This operation does not require authentication. ",
        "operationId": "getSingleReview",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose reviews must be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Review"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Complete a review",
        "description": "The review of the film with ID filmId and issued to the user with ID reviewerId is completed. This operation only allows setting the \"completed\" property to the \"true\" value, and changing the values of the \"reviewDate\", \"rating\", and \"review\" properties. This operation can be performed only by the invited reviewer.",
        "operationId": "updateSingleReview",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be completed",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "The updated Review object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (the user is not the reviewer)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete a review invitation",
        "description": "The review of the film with ID filmId and issued to the user with ID reviewerId is deleted. This operation can only be performed by the owner, and only if the review has not yet been completed by the reviewer.",
        "operationId": "deleteSingleReview",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review invitation must be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation - No Content",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (the user is not the owner)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "409": {
            "description": "Conflict (the review has already been completed)",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Film": {
        "description": "schema for Film data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON data structure describing the film"
          },
          "id": {
            "type": "integer",
            "description": "unique identifier for the film data structure"
          },
          "title": {
            "type": "string",
            "description": "textual title of the film"
          },
          "owner": {
            "type": "integer",
            "description": "unique identifier of the the film data structure owner"
          },
          "private": {
            "type": "boolean",
            "description": "true if only the owner of the film data structure can access it, false otherwise",
            "default": "true"
          },
          "watchDate": {
            "type": "string",
            "format": "date",
            "description": "the date when the film was watched by the owner"
          },
          "rating": {
            "type": "integer",
            "description": "the rating the owner has given to the film",
            "minimum": 1,
            "maximum": 10
          },
          "favorite": {
            "type": "boolean",
            "description": "true if the film is among the favourite ones of the owner, false otherwise ",
            "default": "false"
          },
          "self": {
            "type": "string",
            "format": "uri"
          },
          "reviews": {
            "type": "string",
            "format": "uri",
            "description": "the reviews for the film"
          }
        },
        "required": [
          "title",
          "owner",
          "private"
        ]
      },
      "Films": {
        "description": "schema for Films data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON data structure describing the films"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "description": "total number of pages"
          },
          "currentPage": {
            "type": "integer",
            "minimum": 0,
            "description": "the page included in this object"
          },
          "totalItems": {
            "type": "integer",
            "minimum": 0,
            "description": "the total number of items in the current page"
          },
          "films": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Film"
            },
            "description": "the films of the current page"
          },
          "next": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "totalPages"
        ]
      },
      "Review": {
        "description": "schema for Review data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON data structure describing the film"
          },
          "filmId": {
            "type": "integer",
            "description": "unique identifier of the film for which a review invitation has been issued"
          },
          "reviewerId": {
            "type": "integer",
            "description": "unique identifier of the user who has received the review invitation"
          },
          "completed": {
            "type": "boolean",
            "description": "true if the review has been completed, false otherwise",
            "default": "false"
          },
          "reviewDate": {
            "type": "string",
            "format": "date",
            "description": "the date when the review has been completed by the invited user"
          },
          "rating": {
            "type": "integer",
            "description": "the rating the reviewer has given to the film",
            "minimum": 1,
            "maximum": 10
          },
          "review": {
            "type": "string",
            "description": "a textual description of the review",
            "maxLength": 1000
          },
          "self": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "filmId",
          "reviewerId"
        ]
      },
      "Reviews": {
        "description": "schema for Reviews data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON data structure describing the reviews"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "description": "total number of pages"
          },
          "currentPage": {
            "type": "integer",
            "minimum": 0,
            "description": "the page included in this object"
          },
          "totalItems": {
            "type": "integer",
            "minimum": 0,
            "description": "the total number of items in the current page"
          },
          "films": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "description": "the reviews of the current page"
          },
          "next": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "totalPages"
        ]
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "JWT to identify the authenticated user.",
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}